Graph_Rag/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
neo4j==5.14.0
google-genai==0.3.0
pandas==2.1.1
python-dotenv==1.0.0
